fix : community comment 1차 완료 및 post 수정

+++ src/main/java/com/poten/dive_in/community/comment/service/CommentService.java
@@ -18,6 +18,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -33,15 +34,19 @@ public class CommentService {
     public CommentResponseDTO createComment(CommentRequestDTO requestDTO) {
         Member member = memberRepository.findById(requestDTO.getMemberId()).orElseThrow(() -> new EntityNotFoundException("존재하지 않는 회원입니다."));
         Post post = postRepository.findById(requestDTO.getPostId()).orElseThrow(() -> new EntityNotFoundException("존재하지 않는 글입니다."));
-
+        post.adjustCommentCount(1);
         Comment comment = Comment.builder()
                 .post(post)
                 .member(member)
                 .content(requestDTO.getContent())
                 .cmntClass(0) // 댓글로 설정
                 .orderNumber(0)
+                .likeCount(0)
+                .isActive("Y")
                 .build();
+
         Comment createdComment = commentRepository.save(comment);
+
         createdComment.assignGroupName(createdComment.getId().intValue());
         Comment updatedComment = commentRepository.save(createdComment);
 
@@ -52,7 +57,7 @@ public class CommentService {
     public CommentResponseDTO createReply(Long parentId, CommentRequestDTO requestDTO) {
         Member member = memberRepository.findById(requestDTO.getMemberId()).orElseThrow(() -> new EntityNotFoundException("존재하지 않는 회원입니다."));
         Post post = postRepository.findById(requestDTO.getPostId()).orElseThrow(() -> new EntityNotFoundException("존재하지 않는 글입니다."));
-
+        post.adjustCommentCount(1);
         // 부모 댓글 조회
         Comment parentComment = commentRepository.findById(parentId)
                 .orElseThrow(() -> new EntityNotFoundException("댓글이 존재하지 않습니다."));
@@ -68,6 +73,8 @@ public class CommentService {
                 .groupName(groupName)
                 .orderNumber(orderNumber) // 순서 번호 설정
                 .cmntClass(1) // 대댓글로 설정
+                .likeCount(0)
+                .isActive("Y")
                 .build();
 
         Comment savedReply = commentRepository.save(reply);
@@ -78,7 +85,7 @@ public class CommentService {
     @Transactional
     public void deleteComment(Long commentId, Long memberId) {
         //댓글 존재 확인
-        Comment comment = commentRepository.findByIdWithMember(commentId)
+        Comment comment = commentRepository.findByIdWithMemberAndPost(commentId)
                 .orElseThrow(() -> new EntityNotFoundException("댓글이 존재하지 않습니다."));
 
         // 작성자 확인
@@ -86,11 +93,19 @@ public class CommentService {
             throw new EntityNotFoundException("본인의 댓글만 삭제할 수 있습니다.");
         }
 
-        //대댓글 확인
-        List<Comment> replies = commentRepository.findCommentsByGroupName(comment.getGroupName());
-        for (Comment reply : replies) {
-            commentRepository.delete(reply);
+        Post post = comment.getPost();
+
+        if (comment.getCmntClass() == 0) {
+            List<Comment> replies = commentRepository.findCommentsByGroupName(comment.getGroupName());
+            for (Comment reply : replies) {
+                commentRepository.delete(reply);
+            }
+            int deleteCnt = replies.size();
+            post.adjustCommentCount(-deleteCnt);
+        } else {
+            post.adjustCommentCount(-1);
         }
+        comment.assignPost(post);
         commentRepository.delete(comment);
     }
 
@@ -110,18 +125,29 @@ public class CommentService {
         return CommentResponseDTO.ofEntity(updatedComment);
     }
 
+    @Transactional(readOnly = true)
     public List<PostListResponseDto> getPostsAboutCommentByMemberId(Long memberId, Integer page) {
         int pageSize = 20;
         Pageable pageable = PageRequest.of(page, pageSize);
-        Page<Post> posts = commentRepository.findPostDistinctByMemberId(memberId, pageable);
+        Page<Post> posts = commentRepository.findDistinctPostsByMemberId(memberId, pageable);
         return posts.isEmpty() ? new ArrayList<>() : posts.stream()
                 .map(PostListResponseDto::ofEntity)
                 .collect(Collectors.toList());
     }
 
-    public CommentResponseDTO getCommentById(Long id) {
-        Comment comment = commentRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("존재하지 않는 댓글입니다."));
-        return CommentResponseDTO.ofEntity(comment);
+    @Transactional(readOnly = true)
+    public List<CommentResponseDTO> getCommentById(Long id) {
+        Comment comment = commentRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("존재하지 않는 댓글입니다."));
+
+        List<Comment> comments = commentRepository.findCommentsById(id);
+
+        comments.sort(Comparator.comparing(Comment::getCmntClass)
+                .thenComparing(Comment::getOrderNumber));
+
+        return comments.stream()
+                .map(CommentResponseDTO::ofEntity)
+                .collect(Collectors.toList());
     }
 
 }
